Fix zellij config

+++ zellij/setup.sh
@@ -13,6 +13,8 @@ plugins=(
 for plugin in "${plugins[@]}"; do
   repo="${plugin%/*}"
   file="${plugin##*/}"
-  echo "Installing zellij plugin: $file"
-  curl -L "https://github.com/$repo/releases/latest/download/$file" -o "$PLUGIN_DIR/$file"
+  if [[ ! -f "$PLUGIN_DIR/$file" ]]; then
+    echo "Installing zellij plugin: $file"
+    curl -L "https://github.com/$repo/releases/latest/download/$file" -o "$PLUGIN_DIR/$file"
+  fi
 done
